+++ photorealism-plugin/src/fsr/game_scale.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <windows.h>
+
+namespace photorealism {
+namespace fsr {
+
+void apply_render_scale_to_game(HMODULE proxy_module);
+
+}
+}