+++ photorealism-plugin/src/hooks/present_target.hpp
@@ -5,8 +5,6 @@
 
 namespace photorealism {
 
-void patch_back_buffer_proxy(IDXGISwapChain* swap_chain);
-
 bool present_back_buffer(
     IDXGISwapChain* swap_chain, ID3D11Texture2D** texture);