+++ photorealism-plugin/src/proxy.cpp
@@ -6,6 +6,7 @@
 #include <cwchar>
 
 #include "dinput/input_gate.hpp"
+#include "fsr/game_scale.hpp"
 #include "native_aa/native_aa_config.hpp"
 
 namespace {
@@ -35,6 +36,7 @@ FARPROC real_export(const char* name) {
 
 DWORD WINAPI bootstrap_graphics_proxy(LPVOID) {
     configure_native_aa_for_photorealism(g_proxy_module);
+    photorealism::fsr::apply_render_scale_to_game(g_proxy_module);
     wchar_t sibling_path[MAX_PATH] = {};
     if (g_proxy_module == nullptr ||
         GetModuleFileNameW(g_proxy_module, sibling_path, MAX_PATH) == 0) {