+++ photorealism-plugin/src/steam/capture_pipeline.cpp
@@ -1,4 +1,4 @@
-#include "../hooks/back_buffer_proxy.hpp"
+#include "../hooks/present_target.hpp"
 #include "../runtime.hpp"
 #include "capture_gate.hpp"
 #include "capture_slots.hpp"