+++ photorealism-plugin/tools/build.sh
@@ -31,6 +31,8 @@ common_flags=(
   "${project_dir}/src/dinput/menu_gate.cpp" \
   "${project_dir}/src/hooks/vtable_patch.cpp" \
   "${project_dir}/src/native_aa/apply.cpp" \
+  "${project_dir}/src/fsr/game_scale.cpp" \
+  "${project_dir}/src/fsr/render_scale.cpp" \
   "${project_dir}/src/native_aa/policy.cpp" \
   "${project_dir}/src/native_aa/game_target.cpp" \
   "${project_dir}/src/native_aa/config_file.cpp" \
@@ -57,7 +59,7 @@ common_flags=(
   "${project_dir}/src/hooks/device_probe.cpp" \
   "${project_dir}/src/hooks/vtable_patch.cpp" \
   "${project_dir}/src/hooks/module_names.cpp" \
-  "${project_dir}/src/hooks/back_buffer_proxy.cpp" \
+  "${project_dir}/src/hooks/present_target.cpp" \
   "${project_dir}/src/postprocess/postprocessor.cpp" \
   "${project_dir}/src/postprocess/device_state.cpp" \
   "${project_dir}/src/postprocess/gpu_timer.cpp" \
@@ -125,7 +127,6 @@ common_flags=(
   "${project_dir}/src/overlay/menu_export.cpp" \
   "${project_dir}/src/fsr/render_scale.cpp" \
   "${project_dir}/src/fsr/fsr_telemetry.cpp" \
-  "${project_dir}/src/fsr/proxy_target.cpp" \
   "${project_dir}/src/fsr/upscale_resources.cpp" \
   "${project_dir}/src/fsr/fsr_shaders.cpp" \
   "${project_dir}/src/fsr/upscale_pipeline.cpp" \